+++ app/dashboards/draw_dashboard.rb
@@ -26,7 +26,8 @@ class DrawDashboard < Administrate::BaseDashboard
     last_email_sent: Field::DateTime,
     email_type: Field::String.with_options(searchable: false),
     locking_deadline: Field::DateTime,
-    suite_selection_mode: Field::String.with_options(searchable: false)
+    suite_selection_mode: Field::String.with_options(searchable: false),
+    allow_clipping: Field::Boolean
   }.freeze
 
   # COLLECTION_ATTRIBUTES
@@ -52,6 +53,7 @@ class DrawDashboard < Administrate::BaseDashboard
     locked_sizes
     intent_deadline
     intent_locked
+    allow_clipping
     last_email_sent
     email_type
     suite_selection_mode
@@ -78,6 +80,7 @@ class DrawDashboard < Administrate::BaseDashboard
     groups
     students
     suites
+    allow_clipping
   ).freeze
 
   # Overwrite this method to customize how draws are displayed